Transform wrapper in a HOC
